backend/requirements.txt
python-dotenv
mysql-connector-python
Flask
flask-cors